+++ resources/assets/js/components/presenter/PhotoAlbumPresenter.vue
@@ -25,7 +25,13 @@
 		<carousel ref="carousel" :centerMode="true" :loop="false" :per-page="1" :paginationPosition="'bottom-overlay'" paginationActiveColor="#3897f0" paginationColor="#dbdbdb" class="p-0 m-0">
 			<slide v-for="(img, index) in status.media_attachments" :key="'px-carousel-'+img.id + '-' + index" class="" style="background: #000; display: flex;align-items: center;" :title="img.description">
 
-				<img :class="img.filter_class + ' img-fluid w-100 p-0'" style="" :src="img.url" :alt="altText(img)" onerror="this.onerror=null;this.src='/storage/no-preview.png'">
+				<img
+					:class="img.filter_class + ' img-fluid w-100 p-0'"
+					style=""
+					:src="img.url"
+					:alt="altText(img)"
+					loading="lazy"
+					onerror="this.onerror=null;this.src='/storage/no-preview.png'">
 
 				<p v-if="!status.sensitive && sensitive"
 					@click="status.sensitive = true"
@@ -92,6 +98,8 @@
 </style>
 
 <script type="text/javascript">
+	import BigPicture from 'bigpicture';
+
 	export default {
 		props: ['status'],
 
@@ -113,6 +121,12 @@
 		methods: {
 			toggleContentWarning(status) {
 				this.$emit('togglecw');
+			},
+
+			toggleLightbox(e) {
+				BigPicture({
+					el: e.target
+				})
 			},
 
 			altText(img) {